[cpacker#319] Refactor memgptlog.py to log.py and move a function to log.py
