+++ memgpt/utils.py
@@ -288,11 +288,3 @@ def get_schema_diff(schema_a, schema_b):
     difference = [line for line in difference if line.startswith("+ ") or line.startswith("- ")]
 
     return "".join(difference)
-
-
-def fix_file_path(path):
-    """
-    Converts backslashes to forward slashes in a file path.
-    This is useful for ensuring compatibility in file paths across different systems.
-    """
-    return path.replace("\\", "/")
